phase armor and minor code adjustment
