+++ stats/effects/fu_armoreffects/blistersetbonuseffect.lua
@@ -15,7 +15,7 @@ function init()
 	effect.addStatModifierGroup(armorBonus)	
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 		status.removeEphemeralEffect( "glowpurple" )
