+++ stats/effects/fu_armoreffects/hackersetbonuseffect.lua
@@ -14,7 +14,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	end
