+++ stats/effects/fu_armoreffects/leathersetbonuseffect.lua
@@ -29,7 +29,7 @@ function init()
     end
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	else
