+++ stats/effects/fu_armoreffects/lunarisetbonuseffect.lua
@@ -21,7 +21,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)	 
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	else
