+++ stats/effects/fu_armoreffects/missionarysetbonuseffect.lua
@@ -18,7 +18,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	else
