+++ stats/effects/fu_armoreffects/nomadsetbonuseffect.lua
@@ -35,7 +35,7 @@ function init()
     
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	else
