+++ stats/effects/fu_armoreffects/operativesetbonuseffect.lua
@@ -15,7 +15,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	end
