+++ stats/effects/fu_armoreffects/phasesetbonus.statuseffect
@@ -6,6 +6,5 @@
     
     "phasesetbonus.lua"
   ],
-  "scriptDelta": 50,
-  "animationConfig" : "/stats/effects/damagebonus/damagebonus.animation"
+  "scriptDelta": 50
 }
