+++ stats/effects/fu_armoreffects/phasesetbonuseffect.statuseffect
@@ -1,0 +1,13 @@
+{
+  "name" : "phasesetbonuseffect",
+  "defaultDuration" : 15,
+
+  "scripts" : [
+    
+    "phasesetbonuseffect.lua"
+  ],
+  "scriptDelta": 50,
+  "animationConfig" : "/stats/effects/damagebonus/damagebonus.animation",
+  "label" : "Phase",
+  "icon" : "/interface/statuses/recon.png"
+}
