+++ stats/effects/fu_armoreffects/sanguinesetbonuseffect.lua
@@ -17,7 +17,7 @@ function init()
 	effect.addStatModifierGroup(armorBonus)	
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 		status.removeEphemeralEffect( "regenerationsanguine" )
