+++ stats/effects/fu_armoreffects/shoggothsetbonuseffect.lua
@@ -32,7 +32,7 @@ function init()
 	effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	end
