+++ stats/effects/fu_armoreffects/slimesetbonuseffect.lua
@@ -26,7 +26,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)	
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	else
