+++ stats/effects/fu_armoreffects/stalwartsetbonuseffect.lua
@@ -17,7 +17,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	else
