+++ stats/effects/fu_armoreffects/swashbucklersetbonuseffect.lua
@@ -19,7 +19,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	else
