+++ stats/effects/fu_armoreffects/underworldsetbonuseffect.lua
@@ -18,7 +18,7 @@ function init()
 	effect.addStatModifierGroup(armorBonus)	
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	else
