+++ stats/effects/fu_armoreffects/warangelsetbonuseffect.lua
@@ -27,7 +27,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	end
