+++ stats/effects/fu_armoreffects/xenosetbonuseffect.lua
@@ -14,7 +14,7 @@ function init()
 	effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	end
