+++ stats/effects/fu_armoreffects/xithricitesetbonuseffect.lua
@@ -20,7 +20,7 @@ function init()
 	armorHandle=effect.addStatModifierGroup(armorBonus)
 end
 
-function update()
+function update(dt)
 	if not checkSetWorn(self.setBonusCheck) then
 		effect.expire()
 	end
